Adding comments to each class, and fixing the GUI, also change classes to FINAL.
Adding comments to help explain back and forth what each section works on as i had issues with using the PurchaseLimitExceptions class. Also, decided to use that in the GUI as a limiter on how many items can be purchased by one person at a time in the cart.

Also, decided to make PurchaseLimitException, MusicShopGUI and the Main final classes as they're not inherited by any other class